pyproject.toml
[project]
name = "musicdl-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "musicdl==2.10.2",
    "pycryptodomex>=3.23.0",
    "pyinstaller==6.19.0",
    "pyqt6-stubs==20250824",
    "qthread>=0.1.0",
]
